+++ application/views/User/Dashboard.php
@@ -49,6 +49,10 @@
 					<?php } ?>
 				</select>
 			</div>
+
+			<div id="search-container">
+				<input type="text" class="form-control" aria-label="Search" title="Search. This checks both titles and tags." placeholder="Search" id="search" name="search">
+			</div>
 		</div>
 		<div class="pull-right">
 			<span id="import-status"></span>