+++ public/assets/js/pages/history.js
@@ -1,0 +1,19 @@
+/*global currentPagination:false, totalPagination:false */
+$(function(){
+	"use strict";
+	if(page !== 'history') { return false; }
+
+	$('#history-pagination').pagination({
+		currentPage    : currentPagination,
+
+		pages          : totalPagination,
+		displayedPages : 10,
+
+		hrefTextPrefix : '/user/history/',
+
+		selectOnClick  : false,
+
+		prevText       : '&laquo;',
+		nextText       : '&raquo;'
+	});
+});